fix: iterators i and j were limiting long long to int

+++ img-proc.cpp
@@ -14,7 +14,7 @@ int main(){__
 	long long height, width;cin>>height>>width;
 	if(height <=0 || width <= 0){cout<<"the image is too big!"; return 0;}
 	unsigned long img[width][height];
-	for(int j=0; j<height;j++) for (int i=0; i<width;i++){
+	for(long long j=0; j<height;j++) for (long long i=0; i<width;i++){
 		int r,g,b;cin>>r>>g>>b;
 		img[i][j]=rgb(r,g,b);
 		if(color.find(img[i][j])==color.end()){cout<<"piet does not recognize the color "<< hex << img[i][j] <<" at x= "<< i<< " y= " << j<<endl;return 0;}}
